% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigate_sections.R
\name{previous_sibling}
\alias{previous_sibling}
\title{Jump to previous sibling, climbing when necessary}
\usage{
previous_sibling()
}
\description{
Jump to previous sibling, climbing when necessary
}
